refactor popup into modules
